[Justice Counts] Check null year during bulk upload (Recidiviz/recidiviz-data#40047)
## Description of the change

This PR fixes a bug where Bulk Upload would crash if a row had a null
value in the year column. Now, instead of failing silently, we raise a
user-facing error when this occurs. I've tested this change locally.

## Related issues

Closes Recidiviz/recidiviz-data#39853

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b0c573e3703748194ebf5d03ea8a18405c3ebbcd