[Admin] Allow tasks as a route (Recidiviz/recidiviz-data#40115)
## Description of the change

This PR allows a us to set `tasks` as an allowed route. Tested by
loading the admin panel locally and creating a user with this route.

![image](https://github.com/user-attachments/assets/e02914cf-3b16-45c5-981e-ec7f72ba43b5)

![image](https://github.com/user-attachments/assets/32962a79-5566-41e9-818c-1ea163bd2e28)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of https://github.com/Recidiviz/recidiviz-dashboards/issues/7835

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e365fd9b9b615abe4780d1683d8a3c36e12541e7