[JII Texts] Filter out null phone numbers in jii_to_text (Recidiviz/recidiviz-data#40119)
## Description of the change

This PR filters out people who have `phone_number=NULL` from the product
view that is imported to our DB. This ends up filtering out 9 people.

I preferred to filter them out in the product view and not in the import
logic because we should expect that everyone in the DB is someone who is
textable, thus we should enforce that the `phone_number` is hydrated in
the DB. If we want to be able to text these people, this would require
some upstream work to hunt down these numbers!

Loaded into sandbox with prefix `alexa_`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#40058

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d67c864103eb259842c6da4235aeceb9d3d86f5c