Update V1 states in V2 infra to use state_task_deadline as source for sentence length (Recidiviz/recidiviz-data#39848)
## Description of the change

This is the first step in a series of PRs that are required to migrate
criteria off of supervision and incarceration completion date spans.
This PR injects projected completion dates from
`sentence_deadline_spans_materialized` (view just downstream of state
task deadline) into the new sentencing v2 infra for v1 states. This data
is injected in a v1 version of `sentence_projected_date_sessions`, and
then there is also now a v1-specific version of
`inferred_sentence_group_aggregated_sentence_projected_dates` that
aggregates those sentences to the inferred group level (because the
logic needs to be slightly different for v1 states).

The sentence sessions view dependency graph is now updated with these
new views in yellow (representing views that will be deprecated once
states are migrated off v1).

<img width="1038" alt="Screenshot 2025-03-24 at 2 09 37 PM"
src="https://github.com/user-attachments/assets/40f5796c-81a6-4371-852f-511108dd56ec"
/>

As a result of this change, there will be changes to historical TES
spans for V1 states (because we are allowing the projected dates to
change over time as they now come from state task deadline), but we do
not expect there to be significant changes for sentences currently being
served. There are, however, some changes to dates as a result of
migrating the source to state task deadline, but as explained below,
these are relatively minimal and have no downstream impact on task
eligibility.

**Validation Steps**
I validated the output from these changes in
[this](https://colab.research.google.com/drive/1EtUKhcGTYfNWJY4OSdiGjGNrQK77Kl7T?usp=sharing)
colab notebook. A summary of the validation is as follows:

**1 - Inferred-level projected date comparison**
First, I started upstream and looked for sentence date changes (for
people currently serving sentences) as a result of this PR. **TLDR:
relatively few sentence inferred groups had a date change, and when
there is a date change it can be explained / is expected**. I took all
sentence inferred groups (from `person_projected_date_sessions`) in main
and joined to output from this PR to quantify changes in group projected
dates. For each state, I looked at the % of people that had any date
change, then I broke that up by type of projected date, and then I
categorized the type of date change. As expected, changes are almost
exclusively limited to changes in the
`projected_full_term_release_date_max` field (because this is the
primary field used in sentence deadline spans). In ME, 4.7% of people
currently serving a sentence had a change in their projected completion
date max field, and in all other states this was <1.5%.

Across all states, there were 0.05% (4875) people currently serving a
sentence that had a projected date change as a result of this PR. Of
this 4875, 73% (3573) were a date change from a non-null value to a null
value. For 99.4% of these 3374, this is because the person is currently
serving a life sentence. That is an expected change based on this PR
because that is how dates are calculated in supervision and
incarceration projected completion date spans.

There were 836 cases where a date changed from a null value to a
non-null value, and in all of these cases it is due to dates being
hydrated in state task deadline but not in sentences preprocessed.

And then there were 466 that had an actual date change / discrepancy
between sentence deadline spans and sentences preprocessed. Overall
dates seem to not have changed significantly when there is a discrepancy
between the two sources (median absolute value of number of days
different is 5).

**2 - Criteria Level Validation**
I ran load views to sandbox with the "auto" flag to update all criteria
and task eligibility spans downstream of my changes. Changes to those
that meet criteria are very small as a result of this PR. The criteria
with the largest shift is the `has_active_sentence` criteria in US_TN in
which we now have 0.6% more people that meet this criteria (likely due
to the fact that that criteria uses the `projected_completion_date_max`
field to filter out certain sentences and hydration of that field now
changes slightly).

**3 - Task Level Validation**
There are no changes to any of the TES spans updated as part of this PR!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 04265c540d9a16928ed6f9b55a4060ec406f3424