[US_PA][Workflows] Update to admin ineligible offense logic (Recidiviz/recidiviz-data#39376)
## Description of the change

Ok so PA improved the query that they use to pull criminal history for
us, and a result we received ~8000 new combinations of
statutes/descriptions. My increasingly insane admin ineligible logic
correctly categorized most of these offenses, but mis-categorized ~100
charges. This PR tweaks the admin ineligible offense criterion to
address some of these new errors.

Also! Brenda is working on a validation to catch new combinations of
statutes/descriptions every few weeks, so I will be able to keep an eye
on this in the future.

Validations performed for this PR:
Created a list of charges made admin eligible / ineligible by this PR
[here](https://docs.google.com/spreadsheets/d/19cTRMxcTIvXZojmAn8nL9eWWkRIZntc3C9N6lc-1caI/edit?usp=sharing)
Ran a sandbox to check eligibility changes (prefix sbair) resulting in
2602 -> 2624 reentrants being eligible for admin supervision and 4274 ->
4244 reentrants being eligible for special circ (it makes sense that
special circ eligibility would decrease since if someone is eligible for
admin they will be marked ineligible for special circ).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#39099

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d14e4ffb7a068c66e0cb7397db7926f9055a5e4a