[US_PA][Workflows] Update admin supervision criteria to account for delinquent adjudications (Recidiviz/recidiviz-data#40156)
## Description of the change

Due to the ingest of additional criminal history data, we now have
delinquent adjudications. However, delinquent adjudications only count
towards making someone ineligible for certain offenses (anything that
has an asterisk on [form
402](https://drive.google.com/file/d/1WLLvsGuKPUB4-JycX0KPLv6jqqvnESCi/view?usp=sharing)).
This PR updates the admin supervision offense criterion and case note
helpers to account for these delinquent adjudications.

Note - form 402 technically has an asterisk for the
attempt/conspiracy/solicitation to commit a personal injury crime, but
no asterisks for the actual personal injury crimes.... I want to get a
version of this PR up before the training starts next week so I'm just
going to assume that they messed up and want to include delinquent
adjudications on the actual personal injury crimes too? But I added a
TODO to validate this logic or get them to edit the form for clarity.

Another note! These changes are currently living in a sandbox
(zbrenda_charges, see Recidiviz/recidiviz-data#40152 so make sure you use that if you want to
run anything).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#38994

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9b40cbdc1c59411e1fff675335cda7ddaf80a7b8