[US_TX] Fix how we're handing NULLed out max_termination_dates (Recidiviz/recidiviz-data#40190)
## Description of the change

We are nulling out `Max_Termination_Date`s that are set as magic dates
in the raw data, but that was causing us to accidentally drop periods
for people with those magic `Max_Termination_Date`s . This PR revises it
so that if `Max_Termination_Date` is null in the latest view of the
data, we assume it is a life sentence and we coalesce it temporarily
with 9999-9-9 so those periods don't get dropped. At the end, since we
update end date to be null if it's in the future, the final periods for
these clients will just be open periods.

Note: there are 172 cases where `Max_Termination_Date`s in the raw data
is NULL and in those cases there is no start date in the raw data so
these people's periods would get filtered out anyways. Most of these
clients have a discharged or deceased status.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ae5b2ce5930cd9e9cf276f3760cf692715e8d927