[US_TN] SDS - add 'LATEST' to case-note headers (Recidiviz/recidiviz-data#40176)
## Description of the change

Added the word 'LATEST' to a couple of headers for case-note sections
where we only pull the latest relevant note in the SDS opportunity
record. We already did this for the NCIC check, but I think it would be
good to do the same for the other latest-note sections so that it's more
clear to users why they're only seeing one note in those sections.

To load opportunity record to sandbox:
```
python -m recidiviz.tools.load_views_to_sandbox \
--project_id recidiviz-staging \
--sandbox_dataset_prefix mhilderbran_sds_record manual \
--view_ids_to_load workflows_views.us_tn_suspension_of_direct_supervision_record \
--update_ancestors False --update_descendants False
```

We can check the sandbox view as follows and see that the updated
headers show up for the note sections, as desired:
```
SELECT *
FROM `recidiviz-staging.mhilderbran_sds_record_workflows_views.us_tn_suspension_of_direct_supervision_record_materialized`
```

## Related issues

N/A

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c6690cf6ef20da479f46b048de3faab11fd2adb3