[Testing] Consolidate fixture operations around RawDataFixture (Recidiviz/recidiviz-data#40208)
## Description of the change

This is a small refactor to prepare for Recidiviz/recidiviz-data#38355

It breaks out `RawDataFixture` into its own file and centers the class
around the underlying raw data dependency. This makes it convenient to
check the schemas before reading and writing fixtures. As noted in the
TODOs, I plan on this class being _the way_ to reading and writing
fixtures.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| **Type: Non-breaking refactor** | change addresses some tech debt item
or prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9b301e332188cb6c2cc763fd4c4a68f10fac6d5d