[US_AZ] Remove unused enums  (Recidiviz/recidiviz-data#40162)
## Description of the change

The only LS/RNR assessment currently in use is "LS/RNR
RISK-NEED-RESPONSIVITY". The state was previously using LSI-R
assessments; the “Converted LS/RNR RISK-NEED-RES” is the automated
conversion of historic LSI-R scores to the LS/RNR format. The tests
themselves should exist in both formats, but these are still worth
capturing to have a consistent formatting of both. None of the rest of
these mapped to RISK `StateAssessmentType.LS_RNR` actually appear in the
ingest view results, since they are never referenced in `ofndr_tst`.
This removes them from the mapping for clarity!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 548d7fd69b8ad315fe29ec2fb68c31679e7aaf23