[US_TN] Use new analyst view for completion events (Recidiviz/recidiviz-data#39850)
## Description of the change

Updated our supervision-level-related completion events in TN to use our
latest (and hopefully greatest) counter-backdating logic.

In the process of testing and validating the changes made in Recidiviz/recidiviz-data#38754,
we'd added in a couple more conditions to the logic that helps us decide
when a transition was backdating and whether to adjust the date. We also
now have a new analyst view,
`analyst_data.us_tn_supervision_level_raw_text_sessions_inferred`, that
we can use downstream. I updated our TN-specific completion events to
leverage the latest logic and to use this new view where possible.

To load relevant views to sandbox (manually, at least for the moment):
```
python -m recidiviz.tools.load_views_to_sandbox --project_id recidiviz-staging \
--sandbox_dataset_prefix mhilderbran_us_tn_completion_event_updates manual \
--view_ids_to_load task_eligibility_completion_events_us_tn.supervision_level_downgrade,task_eligibility_completion_events_us_tn.transfer_to_limited_supervision,task_eligibility_completion_events_us_tn.transfer_to_no_contact_parole \
--update_ancestors False --update_descendants False
```

Validation notes:
- We can use the following query (changing the view ID as needed) to
confirm that for each of the completion-event views, we don't see any
changes in the number of completion events per person.
```
WITH main AS (
  SELECT
    state_code,
    person_id,
    COUNT(*) AS n_events_main,
  FROM `recidiviz-staging.task_eligibility_completion_events_us_tn.supervision_level_downgrade_materialized`
  WHERE state_code='US_TN'
  GROUP BY 1, 2
),
sandbox AS (
  SELECT
    state_code,
    person_id,
    COUNT(*) AS n_events_sandbox,
  FROM `recidiviz-staging.mhilderbran_us_tn_completion_event_updates_task_eligibility_completion_events_us_tn.supervision_level_downgrade_materialized`
  GROUP BY 1, 2
)
SELECT *
FROM main
FULL OUTER JOIN sandbox
  USING (state_code, person_id)
WHERE (n_events_main!=sandbox.n_events_sandbox)
  OR (n_events_main IS NULL AND n_events_sandbox IS NOT NULL)
  OR (n_events_main IS NOT NULL AND n_events_sandbox IS NULL)
```
- We can also confidence-check the effects of this PR using the
following query (changing the view ID as needed) to check how the event
dates have changed for each person. NB: this query just compares the
event dates for each person according to the order of the dates
(comparing their earliest date in the main version to their earliest
date in the sandbox, their 4th date in main to their 4th date in the
sandbox, and so on), which means that in this particular query, we can't
definitively link a particular main date to the specific date that
replaced it in the sandbox. We see that very few rows are affected for
each event, which makes sense, and the changes look reasonable.
```
WITH main AS (
  SELECT
    state_code,
    person_id,
    ROW_NUMBER() OVER (PARTITION BY state_code, person_id ORDER BY completion_event_date) AS r,
    completion_event_date AS completion_event_date_main,
  FROM `recidiviz-staging.task_eligibility_completion_events_us_tn.supervision_level_downgrade_materialized`
  WHERE state_code='US_TN'
),
sandbox AS (
  SELECT
    state_code,
    person_id,
    ROW_NUMBER() OVER (PARTITION BY state_code, person_id ORDER BY completion_event_date) AS r,
    completion_event_date AS completion_event_date_sandbox,
  FROM `recidiviz-staging.mhilderbran_us_tn_completion_event_updates_task_eligibility_completion_events_us_tn.supervision_level_downgrade_materialized`
)
SELECT
  state_code,
  person_id,
  completion_event_date_main,
  completion_event_date_sandbox,
  DATE_DIFF(completion_event_date_sandbox, completion_event_date_main, DAY) AS days_diff,
FROM main
FULL OUTER JOIN sandbox
  USING (state_code, person_id, r)
ORDER BY days_diff
```

## Related issues

Relates to Recidiviz/recidiviz-data#39828

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 805c152b0bd5da520697e7f72eeccc33add9a9d4