Fix TX contacts tasks type agnostic joins and reasons json (Recidiviz/recidiviz-data#40239)
## Description of the change

- Fixed the issue where “scheduled” contacts where getting joined to
“unscheduled” contacts as well by changing join from string search to
checking in list
- Maximum General clients have two type agnostic contacts. This messes
up the criteria because we’re not allowed to have two criteria spans at
the same time for the same person, but we do need that for Maximum
General. I removed the quarterly requirement for now until we figure out
how to handle it. Ticket for that here:
https://github.com/Recidiviz/recidiviz-data/issues/40238
  - 1 Home or Field Contact Monthly; and
  - 1 Unscheduled Home or Field Contact Quarterly.
- [[US_TX] Contact cadence
csv](https://docs.google.com/spreadsheets/d/156FzYoM_ULRfrhdx9jCFyM2NYlzTFDPXnFdvfYg5Omo/edit?gid=76070059#gid=76070059)
- Changed the reasons blob to be fully JSON rather than concatenated
text
- Add supervision_level and case_type to reasons blob for easy debugging
- Add contact_types_accepted field to reasons blob for FE

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3eaa9b0f613b788ae186846db4766ccec80ea621