[looker] fix raw data dashboard sort fields (Recidiviz/recidiviz-data#40131)
## Description of the change

when you create a dimension group ex `DOB` it creates fields `DOB_date`,
`DOB_month`, `DOB_year`, etc. so if you try to reference field `DOB` the
field doesn't exist. previously we were using the datetime column name
ex `DOB` to sort dashboard elements. when i added a verification that
all sort fields exist in the field list of a dashboard element, this was
causing UTs to fail so i just commented out the get sort fields logic as
a quick fix.

in the raw data dashboards for datetimes we also create a field like
`DOB__raw` that is the unparsed version of the datetime which is what we
are including the dashboard elements. so, reintroduce the datetime sort
fields using the `__raw` version of the field.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

https://github.com/Recidiviz/recidiviz-data/issues/23292

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 00d6bfbeb37b37aac07d74b57b557b5193361fad