Revert "[US_ME] ineligible for reclass if 30 days from release (v2) (Recidiviz/recidiviz-data#40045)" (Recidiviz/recidiviz-data#40287)
## Description of the change

This reverts commit 57ec9df7fba9463cc8189e85ec7126802a194d35. We
unfortunately forgot to confirm that all validations were passing with
this change that we made. Fortunately, reverting is easy and we will fix
those validations and re-push this fix once Maggie has returned from her
state trip! Yippee!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#40073
Closes Recidiviz/recidiviz-data#40074

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 63133f09f57540b08f6b103cdaeb95e3eafd19fd